Refactor abstract classes to one common class
