+++ src/main/java/org/example/testCases/LoginLogout/SuccessfulLogin.java
@@ -2,11 +2,12 @@ package org.example.testCases.LoginLogout;
 
 import io.github.cdimascio.dotenv.Dotenv;
 import org.example.testCases.Runnable;
+import org.example.testCases.TestCase;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 
-public class SuccessfulLogin extends LogIn implements Runnable {
+public class SuccessfulLogin extends TestCase implements Runnable {
 
     private final Dotenv dotenv = Dotenv.load();
     private static final String URL = "https://jira-auto.codecool.metastage.net/login.jsp?os_destination=%2Fsecure%2FRapidBoard.jspa%3FrapidView%3D3%26projectKey%3DMTP";
