+++ src/main/java/org/example/testCases/Project/BrowseCoalaProject.java
@@ -1,5 +1,6 @@
 package org.example.testCases.Project;
 
+import org.example.testCases.TestCase;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
@@ -8,7 +9,7 @@ import org.openqa.selenium.support.ui.WebDriverWait;
 
 import java.time.Duration;
 
-public class BrowseCoalaProject extends Project implements Runnable {
+public class BrowseCoalaProject extends TestCase implements Runnable {
 
     public BrowseCoalaProject(WebDriver webDriver) {
         super(webDriver);
