+++ src/main/java/org/example/testCases/Project/BrowseProject.java
@@ -1,6 +1,7 @@
 package org.example.testCases.Project;
 
 import org.example.testCases.Runnable;
+import org.example.testCases.TestCase;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
@@ -9,7 +10,7 @@ import org.openqa.selenium.support.ui.WebDriverWait;
 
 import java.time.Duration;
 
-public class BrowseProject extends Project implements Runnable {
+public class BrowseProject extends TestCase implements Runnable {
 
     public BrowseProject(WebDriver webDriver) {
         super(webDriver);
