+++ src/main/java/org/example/testCases/WebDriverProvider.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.testCases;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.WebDriver;
@@ -6,7 +6,7 @@ import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
 
 public class WebDriverProvider {
-    public static final String BROWSER_VERSION = "122";
+    private static final String BROWSER_VERSION = "122";
 
     public static WebDriver setupWebDriver() {
         WebDriverManager.chromedriver().setup();
