+++ src/test/java/org/example/testCases/Issue/SearchIssueTest.java
@@ -1,7 +1,7 @@
 package org.example.testCases.Issue;
 
 import io.github.cdimascio.dotenv.Dotenv;
-import org.example.WebDriverProvider;
+import org.example.testCases.WebDriverProvider;
 import org.example.testCases.LoginLogout.SuccessfulLogin;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
