+++ src/test/java/org/example/testCases/LoginLogout/SuccessfulLogOutTest.java
@@ -1,6 +1,6 @@
 package org.example.testCases.LoginLogout;
 
-import org.example.WebDriverProvider;
+import org.example.testCases.WebDriverProvider;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
